Skeleton class
Skeleton class as an extension of a low poly mesh. The reason to wirte the core as a low poly mesh instead of a network is that, it is just convenient to store a ready made mesh face structure in it. Every time the skeleton gets updated, only the xyz of certain vertices will be need to changes, the structure remain the same.

The network will only be needed when initiate the skeleton, it won't be stored in data dictionary.  The hierachy will be stored as a 'type' value in vertex and edge data.

to_diagram function is embeded so the diagram is just a pure (high poly) mesh.